Add Privacy/Security considerations sections
Closes issue #17

This PR adds "Security Considerations" and "Privacy Considerations"
sections based on the "Self-Review Questionnaire: Security and Privacy."
from the W3C at https://www.w3.org/TR/security-privacy-questionnaire

Below are the answers to the questions:

```
2.1. What information might this feature expose to Web sites or other parties, and for what purposes is that exposure necessary?

When a user performs a screen capture via the [Screen Capture](https://w3c.github.io/mediacapture-screen-share) API, this specification exposes mouse coordinates with the captured surface. It also exposes whether the mouse enters or exits the captured surface.

2.2. Do features in your specification expose the minimum amount of information necessary to enable their intended uses?

No, the mouse coordinates could already be inferred by analyzing the captured video frames and detecting the position of the cursor. However, this approach requires extra code and calculation for web developers and is less convenient than just programmatically accessing coordinates.

2.3. How do the features in your specification deal with personal information, personally-identifiable information (PII), or information derived from them?

No PII can be extracted from mouse coordinates.

2.4. How do the features in your specification deal with sensitive information?

Sensitive information might be inferred on the person capturing their screen by e.g. analyzing the movement of the mouse (speed, tremor, no move at all etc). But as mentioned in (2.2), this is already available from the captured video frames.

2.5. Do the features in your specification introduce new state for an origin that persists across browsing sessions?

No.

2.6. Do the features in your specification expose information about the underlying platform to origins?

These are just mouse coordinates relative to the captured surface which do not depend on the underlying platform. As mentioned in (2.2), this is already available from the captured video frames.

2.7. Does this specification allow an origin to send data to the underlying platform?

No.

2.8. Do features in this specification enable access to device sensors?

No.

2.9. Do features in this specification enable new script execution/loading mechanisms?

No.

2.10 Do features in this specification allow an origin to access other devices?

No.

2.11. Do features in this specification allow an origin some measure of control over a user agent’s native UI?

No.

2.12. What temporary identifiers do the features in this specification create or expose to the web?

None.

2.13. How does this specification distinguish between behavior in first-party and third-party contexts?

The mouse coordinates are only dispatched to the CaptureController object, which lives in the origin from which the method `navigator.mediaDevices.getDisplayMedia()` was called to start the screen capture session. So it is up to the [Screen Capture](https://w3c.github.io/mediacapture-screen-share) to handle extra mitigations.

2.14. How do the features in this specification work in the context of a browser’s Private Browsing or Incognito mode?

This specification behaves the same in normal and private modes, and don't use persistent state.

2.15. Does this specification have both "Security Considerations" and "Privacy Considerations" sections?

Yes once this PR is merged!

2.16. Do features in your specification enable origins to downgrade default security protections?

No.

2.17. How does your feature handle non-"fully active" documents?

The mouse coordinates are only dispatched to the CaptureController object, which lives in the document from which the method `navigator.mediaDevices.getDisplayMedia()` was called to start the screen capture session. When the user navigates away from that document, CaptureController is released and so no more udates are transmitted to the document.

2.18. What should this questionnaire have asked?

This questionnaire should have asked about subtle differences between the coordinates transmitted in CapturedMouseEvents and the information that can be extracted from the captured frames. We elaborate a bit about this topic in the "Security Considerations" and "Privacy Considerations".
```