Linkify "fire an event" concept and use active tense
Paragraph that describes firing of `capturedmousechange` events used but did not
link to the `fire an event` concept. Also, it used passive form, which is
usually bad practice as it somewhat hides the subject of the action (here the
user agent). Also, the spec crawler does not like passive sentences and failed
to extract the event interface.